Fix bug where tree_mask would get incorrectly set in certain cases
As reported in:
https://github.com/paradigmxyz/reth/pull/12080
https://github.com/paradigmxyz/reth/issues/12129

the tree_mask would sometimes get a bit set on it even though the
corresponding child wouldn't appear in the set of updated branch nodes.
This was due to the bit getting incorrectly set for extension nodes
which pointed to updated branch nodes. The solution is a bit of a hack.