feat: retain proofs of non-target nodes in certain edge-cases.
There are specific edge-cases where it's necessary to retain the proofs
for nodes which aren't given in the target set, namely leaf removals
which result in the deletion of a branch, and leaf additions which
result in the creation of a branch.

Documentation of each case is provided in the code at the point it is
handled.

This change will cause more proofs than are strictly necessary to be
retained, because the `target` set we are given does not tell us if
paths are added, updated or removed. This extra work is made up for by
us not needing to fetch the nodes which would otherwise be missing later
down the pipeline, and in benchmarking the overall change comes out very
slightly faster.